[UKET-7] 회원가입/로그인 모듈 마이그레이션
