--- domain/auth-domain/src/main/java/com/uket/domain/auth/config/userid/LoginUserArgumentResolver.java
@@ -1,26 +1,0 @@
-package com.uket.domain.auth.config.userid;
-
-import org.springframework.core.MethodParameter;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(LoginUserId.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-            NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-
-        String userId = SecurityContextHolder.getContext().getAuthentication().getName();
-        return Long.parseLong(userId);
-    }
-}
