--- domain/auth-domain/src/main/java/com/uket/domain/auth/config/userid/LoginUserId.java
@@ -1,12 +1,0 @@
-package com.uket.domain.auth.config.userid;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LoginUserId {
-
-}
