--- domain/auth-domain/src/main/java/com/uket/domain/auth/dto/response/token/OAuth2TokenResponse.java
@@ -1,10 +1,0 @@
-package com.uket.domain.auth.dto.response.token;
-
-public interface OAuth2TokenResponse {
-
-    String getTokenType();
-
-    String getAccessToken();
-
-    String getScope();
-}
