--- domain/auth-domain/src/main/java/com/uket/domain/auth/dto/response/userinfo/GoogleUserInfoResponse.java
@@ -1,35 +1,0 @@
-package com.uket.domain.auth.dto.response.userinfo;
-
-import java.util.Map;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class GoogleUserInfoResponse implements OAuth2UserInfoResponse {
-
-    private final Map<String, Object> attribute;
-
-    @Override
-    public String getProvider() {
-        return "google";
-    }
-
-    @Override
-    public String getProviderId() {
-        return attribute.get("id").toString();
-    }
-
-    @Override
-    public String getEmail() {
-        return attribute.get("email").toString();
-    }
-
-    @Override
-    public String getName() {
-        return attribute.get("name").toString();
-    }
-
-    @Override
-    public String getProfileImage() {
-        return attribute.get("picture").toString();
-    }
-}
