--- domain/auth-domain/src/main/java/com/uket/domain/auth/dto/response/userinfo/OAuth2UserInfoResponse.java
@@ -1,14 +1,0 @@
-package com.uket.domain.auth.dto.response.userinfo;
-
-public interface OAuth2UserInfoResponse {
-
-    String getProvider();
-
-    String getProviderId();
-
-    String getEmail();
-
-    String getName();
-
-    String getProfileImage();
-}
