--- domain/auth-domain/src/main/java/com/uket/domain/auth/exception/AuthException.java
@@ -1,11 +1,0 @@
-package com.uket.domain.auth.exception;
-
-import com.uket.core.exception.BaseException;
-import com.uket.core.exception.ErrorCode;
-
-public class AuthException extends BaseException {
-
-    public AuthException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
