--- domain/auth-domain/src/main/java/com/uket/domain/auth/filter/JwtAuthenticationEntryPoint.java
@@ -1,38 +1,0 @@
-package com.uket.domain.auth.filter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.uket.core.dto.response.ErrorResponse;
-import com.uket.core.exception.ErrorCode;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(
-            HttpServletRequest request, HttpServletResponse response, AuthenticationException authException
-    ) throws IOException {
-        writeErrorResponse(response, authException);
-    }
-
-    private void writeErrorResponse(
-            HttpServletResponse response, AuthenticationException authException
-    ) throws IOException {
-        response.setCharacterEncoding(StandardCharsets.UTF_8.name());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        response.getWriter()
-                .write(objectMapper.writeValueAsString(ErrorResponse.of(ErrorCode.TOKEN_AUTHENTICATION_FAILED)));
-    }
-}
