+++ migration/src/main/java/com/uket/app/auth/config/userid/LoginUserId.java
@@ -1,0 +1,12 @@
+package com.uket.app.auth.config.userid;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginUserId {
+
+}
