+++ migration/src/main/java/com/uket/app/auth/response/token/KakaoTokenResponse.java
@@ -1,0 +1,35 @@
+package com.uket.app.auth.response.token;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.ToString;
+
+public class KakaoTokenResponse implements OAuth2TokenResponse {
+
+    @JsonProperty("token_type")
+    String tokenType;
+    @JsonProperty("access_token")
+    String accessToken;
+    @JsonProperty("expires_in")
+    Long expiredIn;
+    @JsonProperty("refresh_token")
+    String refreshToken;
+    @JsonProperty("refresh_token_expires_in")
+    Long refreshTokenExpiresIn;
+    @JsonProperty("scope")
+    String scope;
+
+    @Override
+    public String getTokenType() {
+        return this.tokenType;
+    }
+
+    @Override
+    public String getAccessToken() {
+        return this.accessToken;
+    }
+
+    @Override
+    public String getScope() {
+        return this.scope;
+    }
+}
