+++ migration/src/main/java/com/uket/app/auth/response/token/OAuth2TokenResponse.java
@@ -1,0 +1,10 @@
+package com.uket.app.auth.response.token;
+
+public interface OAuth2TokenResponse {
+
+    String getTokenType();
+
+    String getAccessToken();
+
+    String getScope();
+}
