+++ migration/src/main/java/com/uket/app/auth/response/userinfo/GoogleUserInfoResponse.java
@@ -1,0 +1,35 @@
+package com.uket.app.auth.response.userinfo;
+
+import java.util.Map;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class GoogleUserInfoResponse implements OAuth2UserInfoResponse {
+
+    private final Map<String, Object> attribute;
+
+    @Override
+    public String getProvider() {
+        return "google";
+    }
+
+    @Override
+    public String getProviderId() {
+        return attribute.get("id").toString();
+    }
+
+    @Override
+    public String getEmail() {
+        return attribute.get("email").toString();
+    }
+
+    @Override
+    public String getName() {
+        return attribute.get("name").toString();
+    }
+
+    @Override
+    public String getProfileImage() {
+        return attribute.get("picture").toString();
+    }
+}
