+++ migration/src/main/java/com/uket/app/auth/response/userinfo/OAuth2UserInfoResponse.java
@@ -1,0 +1,14 @@
+package com.uket.app.auth.response.userinfo;
+
+public interface OAuth2UserInfoResponse {
+
+    String getProvider();
+
+    String getProviderId();
+
+    String getEmail();
+
+    String getName();
+
+    String getProfileImage();
+}
