+++ migration/src/main/java/com/uket/app/domain/user/UserDto.java
@@ -1,0 +1,13 @@
+package com.uket.app.domain.user;
+
+import lombok.Builder;
+
+@Builder
+public record UserDto(
+        Long userId,
+        String name,
+        String role,
+        Boolean isRegistered
+) {
+
+}
